+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 6.0.0
+
+Enhancement: (breaking changes)
+
+- Use python-ipware under the hood
+- Minor behavior changes as python-ipware is more accurate
+- Use 5.0.2 if you need the old behavior
+
 # 5.0.2
 
 Enhancement: